@@ -11,7 +11,7 @@
   "dependencies": {
     "@9gustin/react-notion-render": "^3.11.1",
     "@headlessui/react": "^1.7.11",
-    "@notionhq/client": "^1.0.1",
+    "@notionhq/client": "^1.0.4",
     "@ts-stack/markdown": "^1.4.0",
     "animejs": "^3.2.1",
     "fs": "^0.0.1-security",
@@ -19,9 +19,9 @@
     "gsap": "^3.9.1",
     "kursor": "^0.1.6",
     "marked": "^4.0.12",
-    "next": "^14.1.0",
+    "next": "^14.2.2",
     "next-sitemap": "^3.1.32",
-    "notion-page-to-html": "^1.1.4",
+    "notion-page-to-html": "^1.2.0",
     "react": "^18.2.0",
     "react-animated-cursor": "^2.7.0",
     "react-cursor": "^2.0.15",
