@@ -1,3 +1,4 @@
+"use client";
 import Image from "next/legacy/image";
 import React, { useEffect } from "react";
 import gsap from "gsap";
@@ -25,19 +26,6 @@ const Hero = () => {
         // }
       },
     });
-
-    // timeline.to("#hero-image-container img", {
-    //   translateY: -20,
-    //   duration: 0.5,
-    // });
-    // timeline.to(
-    //   "#hero-image-container svg #clip rect",
-    //   {
-    //     scale: 0.0,
-    //     stagger: 0.1,
-    //   },
-    //   "-=0.5"
-    // );
   }, []);
 
   const animate = () => {
