@@ -46,7 +46,7 @@ const MySetup = () => {
   //   }, []);
 
   return (
-    <div className="container mx-auto relative pt-28 " data-animate-on-scroll>
+    <div className="container mx-auto relative pt-28 ">
       <p className="absolute text-[18vw] lg:text-[15vw] font-extrabold top-10 lg:top-0 -z-10 text-accent opacity-20 dark:opacity-30">My Setup</p>
       <p className="text-2xl lg:text-5xl font-bold text-gray-700 z-10 -mt-4 lg:mt-12 lg:w-2/4 dark:text-white dark:opacity-75">My Setup</p>
 
