@@ -1,11 +1,13 @@
+"use client";
+
 import React, { useEffect } from "react";
 import path from "path";
 import matter from "gray-matter";
 import fs from "fs";
-import Image from "next/legacy/image";
 import Link from "next/link";
 import ScrollTrigger from "gsap/dist/ScrollTrigger";
 import gsap from "gsap";
+import Image from "next/image";
 const RecentWorks = ({ posts }: any) => {
   useEffect(() => {
     gsap.registerPlugin(ScrollTrigger);
@@ -38,21 +40,14 @@ const RecentWorks = ({ posts }: any) => {
     );
   }, []);
 
-  useEffect(() => {
-    // console.log(posts);
-  }, [posts]);
-
-  const recentWorks = [1, 2, 3, 4, 5];
   return (
     <div className="work-container container mx-auto relative pt-28  ">
       <p className="absolute text-[12vw] lg:text-[8vw] font-extrabold top-10 lg:top-0 -z-10 text-accent opacity-20 dark:opacity-30">My Recent Works</p>
-      <h2 data-animate-on-scroll className="text-2xl lg:text-5xl font-bold text-gray-700 z-10 -mt-6 lg:mt-0 lg:w-[70%] dark:text-white ">
-        Here are a few projects I have worked on recently.
-      </h2>
+      <h2 className="text-2xl lg:text-5xl font-bold text-gray-700 z-10 -mt-6 lg:mt-0 lg:w-[70%] dark:text-white ">Here are a few projects I have worked on recently.</h2>
 
       <div className="mt-8 lg:mt-16 mb-12 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-5 gap-5">
         {posts.map((work: any, c: any) => (
-          <Link href={"/works/" + work.slug} passHref key={c}>
+          <Link href={"/projects/" + work.slug} passHref key={c}>
             <div className="work-card shadow-md hover:shadow-2xl opacity-0 scale-50  bg-white rounded-lg aspect-[4/3] cursor-pointer relative">
               <div className="bg-gradient-to-b from-transparent via-transparent rounded-lg to-white absolute inset-0 z-10"></div>
               <h5 className="absolute z-20 bottom-2 left-0 right-0 text-center font-bold">{work.frontmatter.title}</h5>
